Implementacao nova viagem

+++ app/src/main/java/com/example/appfinal/telas/login.kt
@@ -19,7 +19,6 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.dp
-import androidx.lifecycle.viewmodel.compose.viewModel
 import com.example.appfinal.R
 import com.example.appfinal.componentes.CampoSenha
 import model.LoginViewModel
@@ -27,7 +26,7 @@ import model.LoginViewModel
 @Composable
 fun ModLogin(
     onEsqueciSenhaClick: () -> Unit,
-    onLoginClick: () -> Unit,
+    onLoginClick: (Int) -> Unit,
     onNovoUsuario: () -> Unit,
     viewModel: LoginViewModel
 ) {
